requirements.txt
click~=7.1.2
black
ginza
pytest~=5.4.1
tweepy~=3.8.0
pandas~=1.0.3
jsonnet
spacy~=2.2.4
pyyaml